analytics-service/requirements.txt
# MVP リアルタイム分析サービス依存関係

# Web フレームワーク
Flask==2.3.3
gunicorn==21.2.0

# Redis クライアント
redis==5.0.1

# Google Cloud
google-cloud-logging==3.8.0
google-cloud-pubsub==2.18.4

# ユーティリティ
python-dateutil==2.8.2